requirements.txt
asgiref==3.3.1
base58==2.1.0
boto3==1.16.59
botocore==1.19.59
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.9
django-allauth==0.44.0
django-cors-headers==3.6.0
django-filter==2.4.0
django-phonenumber-field==5.0.0
django-rest-auth==0.9.5
django-storages==1.11.1
djangorestframework==3.12.2
drf-extensions==0.6.0
gunicorn==19.9.0
idna==2.10
jmespath==0.10.0
Markdown==3.3.3
more-itertools==8.6.0
oauthlib==3.1.0
Pillow==8.1.0
psycopg2==2.8.6
py-bip39-bindings==0.1.6
py-ed25519-bindings==0.1.2
py-sr25519-bindings==0.1.2
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python-decouple==3.4
python3-openid==3.2.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.3.4
scalecodec==0.10.48
six==1.15.0
sqlparse==0.4.1
substrate-interface==0.11.9
twilio==6.51.0
urllib3==1.26.2
websocket-client==0.57.0
xxhash==2.0.0
